+++ app/src/main/java/woowacourse/payments/newCard/CardCompanySelectBottomSheet.kt
@@ -1,0 +1,34 @@
+package woowacourse.payments.newCard
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ModalBottomSheet
+import androidx.compose.material3.SheetState
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun CardCompanySelectBottomSheet(
+    modalBottomSheetState: SheetState,
+    onClick: (CardCompanyUiModel) -> Unit,
+    isShow: Boolean = true,
+    ) {
+    if (isShow) {
+        ModalBottomSheet(
+            sheetState = modalBottomSheetState,
+            onDismissRequest = { },
+        ) {
+            Box(
+                modifier = Modifier.fillMaxWidth().padding(bottom = 70.dp),
+                contentAlignment = Alignment.Center
+            ) {
+                CardCompanySelectRow(onClick)
+            }
+        }
+    }
+}
