+++ app/src/main/java/woowacourse/payments/newCard/NewCardActivity.kt
@@ -5,27 +5,16 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SheetState
-import androidx.compose.material3.Text
 import androidx.compose.material3.rememberModalBottomSheetState
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -33,12 +22,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import woowacourse.payments.InputMask
 import woowacourse.payments.R
 import woowacourse.payments.domain.Card
@@ -66,7 +51,6 @@ class NewCardActivity : ComponentActivity() {
                     confirmValueChange = { false }
                 )
                 var selectedCardCompany by remember { mutableStateOf(CardCompany.NOT_SELECTED) }
-
 
                 var isShow by remember { mutableStateOf(true) }
                 CardCompanySelectBottomSheet(
@@ -194,80 +178,3 @@ class NewCardActivity : ComponentActivity() {
         }
     }
 }
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CardCompanySelectBottomSheet(
-    modalBottomSheetState: SheetState,
-    onClick: (CardCompanyUiModel) -> Unit,
-    isShow: Boolean = true,
-    ) {
-    if (isShow) {
-        ModalBottomSheet(
-            sheetState = modalBottomSheetState,
-            onDismissRequest = { },
-        ) {
-            Box(
-                modifier = Modifier.fillMaxWidth().padding(bottom = 70.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                CardCompanySelectRow(onClick)
-            }
-        }
-    }
-}
-
-private const val COLUMN_COUNT = 4
-
-@Composable
-fun CardCompanySelectRow(
-    onClick: (CardCompanyUiModel) -> Unit,
-) {
-    val companies = CardCompany.entries
-        .filter { it != CardCompany.NOT_SELECTED }
-        .map { it.toUiModel() }
-
-    FlowRow(
-        modifier = Modifier.padding(vertical = 30.dp, horizontal = 10.dp),
-        horizontalArrangement = Arrangement.spacedBy(12.dp),
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-        maxItemsInEachRow = COLUMN_COUNT
-    ) {
-        companies.forEach { company ->
-            CardCompanyItem(
-                company,
-                onClick = { onClick(company) },
-            )
-        }
-    }
-}
-
-@Composable
-fun CardCompanyItem(
-    cardCompanyUiModel: CardCompanyUiModel,
-    onClick: (CardCompanyUiModel) -> Unit
-) {
-    Box(modifier = Modifier
-        .height(65.dp)
-        .width(80.dp).clickable { onClick(cardCompanyUiModel)},
-    ) {
-        Column(
-            modifier = Modifier
-                .align(alignment = Alignment.Center)
-                .fillMaxHeight(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Icon(
-                painter = painterResource(id = cardCompanyUiModel.iconRes),
-                tint = Color.Unspecified,
-                contentDescription = cardCompanyUiModel.displayName,
-                modifier = Modifier.size(37.dp),
-            )
-            Text(text = cardCompanyUiModel.displayName,
-                fontSize = 16.sp,
-                color = Color(0xFF525252)
-            )
-        }
-    }
-}
-
